# lib/api-zod/src/generated/types/cartItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Lumière dress boutique API
 * OpenAPI spec version: 0.1.0
 */

export interface CartItem {
  id: number;
  productId: number;
  slug: string;
  name: string;
  image: string;
  color: string;
  size: string;
  price: number;
  quantity: number;
  lineTotal: number;
}