# lib/api-zod/src/generated/types/collection.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Lumière dress boutique API
 * OpenAPI spec version: 0.1.0
 */

export interface Collection {
  slug: string;
  name: string;
  description: string;
  coverImage: string;
  productCount: number;
}