# lib/api-zod/src/generated/types/facets.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Lumière dress boutique API
 * OpenAPI spec version: 0.1.0
 */
import type { FacetsCategoriesItem } from "./facetsCategoriesItem";
import type { FacetsCollectionsItem } from "./facetsCollectionsItem";
import type { FacetsColorsItem } from "./facetsColorsItem";
import type { FacetsPriceRange } from "./facetsPriceRange";
import type { FacetsSizesItem } from "./facetsSizesItem";

export interface Facets {
  categories: FacetsCategoriesItem[];
  collections: FacetsCollectionsItem[];
  colors: FacetsColorsItem[];
  sizes: FacetsSizesItem[];
  priceRange: FacetsPriceRange;
}